Partially revert "1024 - Reorganize”, use different interface name
