+++ server/libs/platform/platform-component/platform-component-registry/platform-component-registry-service/src/main/java/com/bytechef/platform/component/registry/factory/config/ComponentHandlerServiceLoaderFactory.java
@@ -1,0 +1,34 @@
+/*
+ * Copyright 2023-present ByteChef Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.bytechef.platform.component.registry.factory.config;
+
+import com.bytechef.component.ComponentHandler;
+import com.bytechef.platform.component.factory.ComponentHandlerFactory;
+import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
+import java.util.List;
+
+/**
+ * @author Ivica Cardic
+ */
+@SuppressFBWarnings("EI")
+public record ComponentHandlerServiceLoaderFactory(
+    List<? extends ComponentHandler> componentHandlers) implements ComponentHandlerFactory {
+
+    public List<? extends ComponentHandler> getComponentHandlers() {
+        return componentHandlers;
+    }
+}
